+++ packages/dialect-react-ui/components/NotificationsButton/index.tsx
@@ -34,6 +34,7 @@ export type PropTypes = {
   notifications: NotificationType[];
   channels?: Channel[];
   onBackClick?: () => void;
+  pollingInterval?: number;
 };
 
 export function useOutsideAlerter(
@@ -168,7 +169,10 @@ export default function NotificationsButton({
   return (
     <div className="dialect">
       <ApiProvider dapp={props.publicKey.toBase58()}>
-        <DialectProvider publicKey={props.publicKey}>
+        <DialectProvider
+          pollingInterval={props.pollingInterval}
+          publicKey={props.publicKey}
+        >
           <ThemeProvider theme={theme} variables={variables}>
             <WrappedNotificationsButton channels={channels} {...props} />
           </ThemeProvider>