+++ packages/dialect-react-ui/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added logic to notfication ui to display new notification badge
 - fix: open thread if user entered address with whom thread already exists
+- fix: do not fetch handles and avatars twice
 
 
 ## [0.8.9] - 2022-06-07